\name{model.frame.rpart}
\alias{model.frame.rpart}
\title{ Model Frame Method for rpart }
\description{

  A model.frame method for rpart objects.

}
\usage{
\method{model.frame}{rpart}(formula, \dots)
}
\arguments{
  \item{formula}{ an object of class \code{\link{rpart}}.}
  \item{\dots}{ additional arguments.}
}
\details{
  A \code{\link{model.frame}} method for \code{\link{rpart}} objects.
}
\value{
  A model frame.
}
\keyword{tree}
