\name{varimp}
\alias{varimp}
\alias{varimp.constparty}
\alias{varimp.cforest}
\title{ Variable Importance }
\description{
    Standard and conditional variable importance for `cforest', following the permutation
    principle of the `mean decrease in accuracy' importance in `randomForest'.
}
\usage{
varimp(object, nperm = 1L, ...)
\method{varimp}{constparty}(object, nperm = 1L, 
       risk = c("loglik", "misclassification"), conditions = NULL,
       mincriterion = 0, ...)
\method{varimp}{cforest}(object, nperm = 1L, 
       OOB = TRUE, risk = c("loglik", "misclassification"),
       conditional = FALSE, threshold = .2, applyfun = NULL, 
       cores = NULL, ...)
}
\arguments{
  \item{object}{ an object as returned by \code{cforest}.}
  \item{mincriterion}{ the value of the test statistic or 1 - p-value that
                       must be exceeded in order to include a split in the 
                       computation of the importance. The default \code{mincriterion = 0}
                       guarantees that all splits are included.}
  \item{conditional}{ a logical determining whether unconditional or conditional 
                      computation of the importance is performed. }
  \item{threshold}{ the value of the test statistic or 1 - p-value of the association 
                    between the variable of interest and a covariate that must be 
                    exceeded inorder to include the covariate in the conditioning 
                    scheme for the variable of interest (only relevant if 
                    \code{conditional = TRUE}). }
  \item{nperm}{ the number of permutations performed.}
  \item{OOB}{ a logical determining whether the importance is computed from the out-of-bag 
              sample or the learning sample (not suggested).}
  \item{risk}{ a character determining the risk to be evaluated.}
  \item{conditions}{ a list of conditions. }
  \item{applyfun}{an optional \code{\link[base]{lapply}}-style function with arguments
                  \code{function(X, FUN, \dots)}. It is used for computing the variable importances for each tree. 
                  The default is to use the basic \code{lapply}
                  function unless the \code{cores} argument is specified (see below).
                  Extra care is needed to ensure correct seeds are used in
                  the parallel runs (\code{RNGkind("L'Ecuyer-CMRG")} for
                  example).}
  \item{cores}{numeric. If set to an integer the \code{applyfun} is set to
               \code{\link[parallel]{mclapply}} with the desired number of \code{cores}.}
  \item{\dots}{additional arguments, not used.}
}
\details{

  NEEDS UPDATE

  Function \code{varimp} can be used to compute variable importance measures
  similar to those computed by \code{\link[randomForest]{importance}}. Besides the
  standard version, a conditional version is available, that adjusts for correlations between
  predictor variables. 
  
  If \code{conditional = TRUE}, the importance of each variable is computed by permuting 
  within a grid defined by the covariates that are associated  (with 1 - p-value 
  greater than \code{threshold}) to the variable of interest.
  The resulting variable importance score is conditional in the sense of beta coefficients in   
  regression models, but represents the effect of a variable in both main effects and interactions.
  See \bibcitet{partykit::Strobl+Boulesteix+Kneib:2008} for details.

  Note, however, that all random forest results are subject to random variation. Thus, before
  interpreting the importance ranking, check whether the same ranking is achieved with a
  different random seed -- or otherwise increase the number of trees \code{ntree} in 
  \code{\link{ctree_control}}.

  Note that in the presence of missings in the predictor variables the procedure
  described in \bibcitet{partykit::Hapfelmeier_Hothorn_Ulm_Strobl_2014} is performed.

}
\value{
  A vector of `mean decrease in accuracy' importance scores.
}

\references{\bibshow{*}}

\examples{
    
   set.seed(290875)
   if (requireNamespace("party")) {
       data("readingSkills", package = "party")
       readingSkills.cf <- cforest(score ~ ., data = readingSkills, 
                                   mtry = 2, ntree = 50)

       # standard importance
       varimp(readingSkills.cf)

       # conditional importance, may take a while...
       varimp(readingSkills.cf, conditional = TRUE)
   }
}
\keyword{tree}
